pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "Hybrid RL+LLM email triage environment for OpenEnv hackathon"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "groq>=0.9.0",
    "python-dotenv>=0.21.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"
